+++ test/specs/openapi/test_schemas.py
@@ -1,4 +1,7 @@
+from functools import partial
+
 import pytest
+from hypothesis import find
 
 import schemathesis
 from schemathesis.models import APIOperation
@@ -25,3 +28,38 @@ def test_get_operation_via_remote_reference(openapi_version, schema_url):
             "required": True,
             "schema": {"type": "integer"},
         }
+
+
+SINGLE_METHOD_PATHS = {
+    "/test-2": {"get": {"responses": {"200": {"description": "OK"}}}},
+}
+TWO_METHOD_PATHS = {
+    "/test": {
+        "get": {"responses": {"200": {"description": "OK"}}},
+        "post": {"responses": {"200": {"description": "OK"}}},
+    },
+}
+
+
+def matches_operation(case, operation):
+    return operation.method.upper() == case.method.upper() and operation.full_path == case.full_path
+
+
+def test_path_as_strategy(empty_open_api_3_schema):
+    empty_open_api_3_schema["paths"] = TWO_METHOD_PATHS
+    schema = schemathesis.from_dict(empty_open_api_3_schema)
+    operations = schema["/test"]
+    strategy = operations.as_strategy()
+    for operation in operations.values():
+        # All fields should be possible to generate
+        find(strategy, partial(matches_operation, operation=operation))
+
+
+def test_schema_as_strategy(empty_open_api_3_schema):
+    empty_open_api_3_schema["paths"] = {**SINGLE_METHOD_PATHS, **TWO_METHOD_PATHS}
+    schema = schemathesis.from_dict(empty_open_api_3_schema)
+    strategy = schema.as_strategy()
+    for operations in schema.values():
+        for operation in operations.values():
+            # All operations should be possible to generate
+            find(strategy, partial(matches_operation, operation=operation))